Refactoring to make it easier to understand
