+++ treeple/tree/__init__.py
@@ -7,7 +7,6 @@ from .._lib.sklearn.tree import (
 from ._classes import (
     ExtraObliqueDecisionTreeClassifier,
     ExtraObliqueDecisionTreeRegressor,
-    KernelDecisionTreeClassifier,
     ObliqueDecisionTreeClassifier,
     ObliqueDecisionTreeRegressor,
     PatchObliqueDecisionTreeClassifier,
@@ -16,6 +15,7 @@ from ._classes import (
     UnsupervisedObliqueDecisionTree,
 )
 from ._honest_tree import HonestTreeClassifier
+from ._kernel import KernelDecisionTreeClassifier
 from ._multiview import MultiViewDecisionTreeClassifier
 from ._neighbors import compute_forest_similarity_matrix
